pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-trust-safety"
version = "0.1.0"
description = "A real-world Trust & Safety environment for OpenEnv."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai"
]

[project.scripts]
server = "server.app:main"